% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_builders.R
\name{apply_axis_controls}
\alias{apply_axis_controls}
\title{Apply optional y-axis controls specified in plotControl()}
\usage{
apply_axis_controls(p, arg)
}
\description{
Apply optional y-axis controls specified in plotControl()
}
\keyword{internal}
