% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC.R
\name{CCC}
\alias{CCC}
\title{Internal Function to Compute the Sampled Concordance
  Correlation Values.}
\value{
No return value, called for side effects
}
\description{
This is an internally called functions used to compute
  the sampled concordance correlation values.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
