% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lcc.R
\name{CCC_lin}
\alias{CCC_lin}
\title{Internal Function to Estimate the Sampled Concordance
  Correlation Coefficient.}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to estimate
  the sampled concordance correlation coefficient.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
