% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciBuilder.R
\name{ciBuilder}
\alias{ciBuilder}
\title{Internal Function to Prepare the \code{ciCompute} Function.}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to prepare
  the \code{\link[lcc]{ciCompute}} function.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
