% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccBootstrap.R
\name{dataBootstrap}
\alias{dataBootstrap}
\title{Internal Functions to Generate Bootstrap Samples Based on
  Dataset.}
\value{
No return value, called for side effects
}
\description{
This is an internally called functions used to generate
  bootstrap samples.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
