% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_wraper_lcc_lpc_la.R
\name{.compute_LA}
\alias{.compute_LA}
\title{Internal helper to compute LA list for one diffbeta}
\usage{
.compute_LA(pre, diffbeta)
}
\description{
Internal helper to compute LA list for one diffbeta
}
\keyword{internal}
