% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_wraper_lcc_lpc_la.R
\name{.compute_LCC}
\alias{.compute_LCC}
\title{Internal helper to compute LCC list for one diffbeta}
\usage{
.compute_LCC(pre, diffbeta)
}
\description{
Internal helper to compute LCC list for one diffbeta
}
\keyword{internal}
