% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_wraper_lcc_lpc_la.R
\name{.compute_LPC}
\alias{.compute_LPC}
\title{Internal helper to compute LPC list (does not depend on diffbeta)}
\usage{
.compute_LPC(pre)
}
\description{
Internal helper to compute LPC list (does not depend on diffbeta)
}
\keyword{internal}
