% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{.lcc_default_theme}
\alias{.lcc_default_theme}
\title{Internal base theme used by all ggplot-based summaries}
\usage{
.lcc_default_theme()
}
\description{
Internal base theme used by all ggplot-based summaries
}
\keyword{Internal}
