% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_components.R
\name{extract_random_effects_cov}
\alias{extract_random_effects_cov}
\title{Internal Function to Extract Variance Components Estimates.}
\description{
This is an internally called function used to extract
  variance components estimate of \eqn{\Sigma} matrix based on
  specified structure.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br} and Rafael de Andrade Moral,
  \email{rafael_moral@yahoo.com.br}
}
\keyword{internal}
