% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedBuilder.R
\name{fittedBuilder}
\alias{fittedBuilder}
\title{Internal Function to Build Fitted Values for
  \code{lcc} Objects}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to build
  fitted values.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
