% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDelta.R
\name{getDelta}
\alias{getDelta}
\title{Internal Function to Extract Variance Components Estimates.}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to extract
  variance components estimate of \eqn{\Sigma} matrix based on
  specified structure.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br} and Rafael de Andrade Moral,
  \email{rafael_moral@yahoo.com.br}
}
\keyword{internal}
