% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{init}
\alias{init}
\title{Internal Function to Prepare \code{\link[lcc]{lccModel}}
  Function}
\value{
A named list with resolved arguments used internally by lccModel
}
\description{
This is an internally called function used to verify
  the specification of variance-covariance matrices and likelihood
  based method.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
