% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{is.lcc}
\alias{is.lcc}
\title{Reports whether x is a lcc object}
\usage{
is.lcc(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Returns if the object belongs to lcc class
}
\description{
Reports whether x is a lcc object
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
