% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccBootstrap.R
\name{laBootstrap}
\alias{laBootstrap}
\title{Internal functions to generate longitudinal accuracy samples.}
\value{
No return value, called for side effects
}
\description{
This is an internally called functions used to generate
  longitudinal caccuracy samples.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
