% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laWrapper.R
\name{laWrapper}
\alias{laWrapper}
\title{Internal Function to Prepare the \code{laBuilder} Function}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to prepare the \code{\link[lcc]{laBuilder}} function.
}
\details{
returns a vector or list containing the longitudinal accuracy estimates.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
