% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc_intervals.R
\name{lcc_intervals}
\alias{lcc_intervals}
\title{Internal Functions to Compute the Non-Parametric Confidence
  Intervals for LCC.}
\value{
No return value, called for side effects
}
\description{
This is an internally called functions used to compute
  the non-parametric confidence intervals for LCC.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
