% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc_internal.R
\name{lccInternal}
\alias{lccInternal}
\title{Internal Function to Prepare \code{lcc} Objects}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to prepare
  \code{lcc} objects for calculate the longitudinal concordance
  correlation, longitudinal Pearson correlation, longitudinal bias
  corrector, and plotting
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br} and Rafael de Andrade Moral,
  \email{rafael_moral@yahoo.com.br}
}
\keyword{internal}
