% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccModel.R
\name{lccModel}
\alias{lccModel}
\title{Internal Function to Fits a Linear Mixed-Effects Model in the
  Formulation Described in Laird and Ware (1982).}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to fits a
  linear mixed-effects model; see \code{\link[nlme]{lme}}.
}
\references{
Laird, N.M. and Ware, J.H. (1982) Random-Effects Models
  for Longitudinal Data, \emph{Biometrics}, 38, 963–974.

Pinheiro, J.C. and Bates., D.M. (1996) Unconstrained Parametrizations
for Variance-Covariance Matrices, \emph{Statistics and Computing},
6, 289–296.

Pinheiro, J.C., and Bates, D.M. (2000) Mixed-Effects Models in S and
S-PLUS, \emph{Springer}.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
