% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccSummary.R
\name{lccSummary}
\alias{lccSummary}
\title{Internal Function to Summarize Fitted and Sampled Values for
  \code{lcc} Objects}
\value{
No return value, called for side effects
}
\description{
This is an internally called function used to summarize
  fitted and sampled values, and the concordance correlation
  coefficient between them for \code{lcc} objects.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
