% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccWrapper.R
\name{lccWrapper}
\alias{lccWrapper}
\title{Internal Function to Prepare the \code{lccBuilder}
  Function}
\value{
returns a vector or list containing the longitudinal
  concordance correlation estimates.
}
\description{
This is an internally called function used to prepare
  the \code{\link[lcc]{lccBuilder}} function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
