% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcBuilder.R
\name{lpcBuilder}
\alias{lpcBuilder}
\title{Internal Function to Estimate the Longitudinal Pearson
  Correlation.}
\value{
returns a vector or list containing the longitudinal
  Pearson correlation estimates.
}
\description{
This is an internally called function used to estimate
  the longitudinal Pearson correlation (LPC).
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
