% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcWrapper.R
\name{lpcWrapper}
\alias{lpcWrapper}
\title{Internal Function to Prepare the \code{lpcBuilder}
  Function}
\value{
returns a vector or list containing the longitudinal
  Pearson correlation estimates.
}
\description{
This is an internally called function used toprepare
  the \code{\link[lcc]{lpcBuilder}} function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
