% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpc.R
\name{Pearson}
\alias{Pearson}
\title{Internal Function to Estimate the Sampled Pearson
  Correlation.}
\value{
No return value, called for side effects
}
\description{
This is an internally called functions used to estimate
  the sampled Pearson correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
