% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpc.R
\name{plot_lpc}
\alias{plot_lpc}
\title{Internal function to prepare the
  \code{\link[lcc]{plotBuilder_lpc}} function.}
\description{
This is an internally called functions used to prepare
  the \code{\link[lcc]{plotBuilder_lpc}} function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
