% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LCC.R
\name{plot_screening}
\alias{plot_screening}
\title{Plot the screening}
\usage{
plot_screening(list_screening, title = "")
}
\arguments{
\item{list_screening}{an object from LCC (either $Area or $Count)}

\item{title}{\code{character} a title for the plot}
}
\description{
Typically after an LCC() step, plots the fire detected by the three methods
}
\examples{
\dontrun{
# see LCC examples.
}
}

