% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rename.vars}
\alias{rename.vars}
\title{Internal Function to Prepare \code{\link[lcc]{lccModel}}
  Function}
\description{
This is an internally called function used to verify
  the specification of variance-covariance matrices and likelihood
  based method.
}
\author{
Code by Don MacQueen
}
\keyword{internal}
