% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBuilder.R
\name{dataBuilder}
\alias{dataBuilder}
\title{Internal Function to Prepare the Dataset for \code{lcc}
Objects}
\description{
This is an internally called function used to prepare
  the dataset for \code{lcc} objects
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
