% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laBuilder.R
\name{laBuilder}
\alias{laBuilder}
\title{Internal Function to Estimate the Longitudinal Accuracy.}
\description{
This is an internally called function used to estimate
  the longitudinal accuracy (LA).
}
\details{
returns a vector or list containing the longitudinal
  accuracy estimates.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
