% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_wraper_lcc_lpc_la.R
\name{laBuilder}
\alias{laBuilder}
\title{Internal Function to Estimate the Longitudinal Accuracy (Bias Corrector)}
\usage{
laBuilder(G, diffbeta, tk, q_r, q_f, g, sig2_epsilon, delta, deltal, model)
}
\description{
Thin wrapper around .precompute_longitudinal() and .compute_LA().
}
\keyword{internal}
