% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LCC.R
\name{LCC}
\alias{LCC}
\title{Performs Large Charcoal Count method}
\usage{
LCC(path = getwd(), write.csv = FALSE, min.charnb = 10,
  wdw.increment = 400, wdw.width = 1000, FUN = median)
}
\arguments{
\item{path}{the path where to find the required files. \code{getwd()} by default}

\item{write.csv}{\code{logical} whether to write \code{.csv} files}

\item{min.charnb}{\code{numeric} minimum number of charcoals in a peak to retain it (10 by default)}

\item{wdw.increment}{\code{numeric} the increment for sliding the window (400 by default)}

\item{wdw.width}{\code{numeric} the width of the moving window (1000 by default)}

\item{FUN}{\code{function} returning a scalar, to apply to the charcoal retrieved within
a moving window. (\code{median} by default)}
}
\value{
a list with two components : \code{$Area} and \code{$Count}, each of them
containing three components: \code{$raw} (raw results from CharAnalysis),
\code{$sc} screened results with ARCO and \code{$LCC} screening with the LCC methods.

\code{$Area} and \code{$Count} can be passed directly to \link{plot_screening}.
}
\description{
Performs Large Charcoal Count method
}
\details{
This functions needs several input data, in 'path', by default the current working directory.
Have a look to those provided available theregi: 'http://www.github/vbonhomme/LCC/data.zip'

I. Raw data:

I.1. \strong{Seedle.csv}: a .csv table with charcoal-particle areas.
Must have three columns and as many rows as the number of observations. These columns
must be labelled as follows:
\itemize{
\item 'Depth': depth of samples
\item 'SdlCounts': number of charcoal particles in each sample
\item 'SdlArea': charcoal-particle areas
}

I.2. \strong{Smpl.csv}: a .csv table with charcoal counts and charcoal areas.
Must have four columns and as many rows as the number of samples. These columns
must be labelled as follows:
\itemize{
\item 'Depth': depth of samples
\item 'Age_calBP': age estimate of samples
\item 'SmplCount': number of charcoal particles in each sample
\item 'SmplArea': cumulative charcoal area in each sample
}

II. Fire episode reconstructions:
These are output files from the
\itemize{

\item CharAnalysis program by Higuera et al. (2009), which
is freely available there \url{http://sites.google.com/site/charanalysis/}.
\item ARCO by Finsinger et al. (2014) which is freely available there \url{https://github.com/wfinsinger/ARCO}
}

In file names below, A stands for Area; C for counts; sc for screening with ARCO.

\itemize{
\item 'FireA.csv': CharAnalysis output table from analysis of CHARCOAL AREAS
\item 'FireAsc.csv': CharAnalysis output table from analysis of CHARCOAL AREAS and
 WITH ARCO, with those parameters \code{n.boot=10000}, \code{thresh.prob=0.95},
\code{win.width=1000} and \code{breakage=FALSE}. Also, \code{minCountP} must be < 1.0
(e.g. 0.05 as in Higuera et al. 2009)
\item 'FireC.csv': CharAnalysis output table from analysis of CHARCOAL COUNTS
\item 'FireCsc.csv': CharAnalysis output table from analysis of CHARCOAL COUNTS and
with minimum count test, i.e. with CharAnalysis parameter \code{minCountP < 1.0}, e.g. 0.05 as in Higuera et al. (2009).
 }

 Note that relevant information from all CharAnalysis is done via column number because
 columns names are poorly formatted. Thus, files must have columns in the expected order, specifically:
 \itemize{
 \item Column 2:  'age Top_i', the age at the top of the interpolated sample
 \item Column 8:  'char Peak', the Cpeak component of CHAR
 \item Column 12: 'thresh FinalPos', the threshold used for peak identification
 \item Column 19: 'peaks Final', the boolean series representing identified peaks.
 }
}
\examples{
\dontrun{
# If you have all required files in your working directory.
# You can find a .zip with example data there: 'http://www.github/vbonhomme/LCC/data.zip'
results <- LCC()
plot_screening(results$Area, "Area")
plot_screening(results$Count, "Count")
# Also, you should now have FireA, FireAsc, FireC, FireCsc, Seedle and Smpl
available in your environment.
# if write.csv=TRUE, you should also have .csv files in your working directory.
}
}
\references{
The method is detailed in a submitted (2015) paper by Remy et al.

Papers cited in the 'Details' section:
\itemize{
\item Finsinger, W., R. Kelly, J. Fevre, and E.K. Magyari (2014). A guide to screening
    charcoal peaks in macrocharcoal-area records for fire episode reconstructions.
    \emph{The Holocene} 24(8):1002-108.
    It can be found there \url{http://hol.sagepub.com/content/24/8/1002}

\item Higuera, P.E., L.B. Brubaker, P.M. Anderson, F.S. Hu, and T.A. Brown (2009)
    Vegetation mediated the impacts of postglacial climate change on fire regimes in
    the south-central Brooks Range, Alaska. Ecological Monographs 79(2) 201:219.
   It can be foudn there \url{http://www.esajournals.org/doi/abs/10.1890/07-2019.1}
   }
}

