% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccBuilder.R
\name{lccBuilder}
\alias{lccBuilder}
\title{Internal Function to Estimate the Longitudinal Concordance
  Correlation.}
\description{
This is an internally called function used to estimate
  the longitudinal concordance correlation (LCC).
}
\details{
returns a vector or list containing the longitudinal
  concordance correlation estimates.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
