% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_metrics.R
\name{lccBuilder}
\alias{lccBuilder}
\title{Internal Function to Estimate the Longitudinal Concordance Correlation}
\usage{
lccBuilder(G, diffbeta, tk, q_r, q_f, g, sig2_epsilon, delta, deltal, model)
}
\description{
Thin wrapper around .precompute_longitudinal() and .compute_LCC().
}
\keyword{internal}
