% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcc_workflow.R
\name{lccInternal}
\alias{lccInternal}
\title{Internal Function to Prepare \code{lcc} Objects}
\arguments{
\item{boot.scheme}{character bootstrap scheme passed from \code{lcc};
defaults to \code{"np_case"}.}

\item{ci.method}{character CI method passed from \code{lcc};
defaults to \code{"normal"}.}
}
\description{
This is an internally called function used to prepare
  \code{lcc} objects for calculate the longitudinal concordance
  correlation, longitudinal Pearson correlation, longitudinal bias
  corrector, and plotting
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br} and Rafael de Andrade Moral,
  \email{rafael_moral@yahoo.com.br}
}
\keyword{internal}
