% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccModel.R
\name{lccModel}
\alias{lccModel}
\title{Internal function to fit a linear mixed-effects model}
\description{
Internal helper to fit a linear mixed-effects model,
  following the formulation described in Laird and Ware (1982).
  See \code{\link[nlme]{lme}} for details.
}
\references{
Laird, N. M.; Ware, J. H. (1982). Random-effects models
  for longitudinal data. \emph{Biometrics}, 38, 963–974.

Pinheiro, J. C.; Bates, D. M. (1996). Unconstrained
  parametrizations for variance–covariance matrices.
  \emph{Statistics and Computing}, 6, 289–296.

Pinheiro, J. C.; Bates, D. M. (2000). Mixed-effects
  models in S and S-PLUS. Springer.
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
