% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccSummary.R
\name{lccSummary}
\alias{lccSummary}
\title{Internal Function to Summarize Fitted and Sampled Values for \code{lcc} Objects}
\description{
Internally called function for summarizing fitted and sampled values, and the 
  concordance correlation coefficient between them for \code{lcc} objects.
}
\details{
Returns a summary of fitted and sampled values and their concordance correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
