% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccBootstrap.R
\name{lpcBootstrap}
\alias{lpcBootstrap}
\title{Internal functions to generate longitudinal Pearson correlation samples.}
\description{
This is an internally called functions used to generate longitudinal Pearson correlation samples.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
