% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_metrics.R
\name{lpcBuilder}
\alias{lpcBuilder}
\title{Internal Function to Estimate the Longitudinal Pearson Correlation}
\usage{
lpcBuilder(G, tk, q_r, q_f, g, sig2_epsilon, delta, deltal, model)
}
\description{
Thin wrapper around .precompute_longitudinal() and .compute_LPC().
}
\keyword{internal}
