% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{Pearson}
\alias{Pearson}
\title{Estimate Sampled Pearson Correlation}
\description{
Internally called function to estimate the sampled Pearson correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
