% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_la.R
\name{plot_la}
\alias{plot_la}
\title{Internal function to prepare the \code{plotBuilder_la} function.}
\description{
This is an internally called functions used to prepare the \code{\link[lcc]{plotBuilder_la}} function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
