% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_la}
\alias{plot_la}
\title{Internal Function to Prepare the \code{plotBuilder_la} Function}
\description{
This function is internally called to prepare
  the \code{\link[lcc]{plotBuilder_la}} function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
