% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_builders.R
\name{plot_lpc}
\alias{plot_lpc}
\title{Prepare Plot for LPC Function}
\description{
Internally called function to prepare data for the `plotBuilder_lpc` function.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
