% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBuilder_la.R
\name{plotBuilder_la}
\alias{plotBuilder_la}
\title{Internal Function to Produces a Longitudinal Accuracy Plot.}
\description{
This is an internally called function used to produces
  a longitudinal accuracy plot from fitted ans sampled values with
  or not non-parametric confidence intervals.
}
\details{
returns a inital plot for the longitudinal accuracy
  correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@usp.br}
}
\keyword{internal}
