% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBuilder_la.R
\name{plotBuilder_la}
\alias{plotBuilder_la}
\title{Internal Function to Produces a Longitudinal Accuracy Plot.}
\description{
Produces a longitudinal accuracy plot from fitted and sampled values 
  with optional non-parametric confidence intervals.
}
\details{
Returns an initial plot for the longitudinal accuracy correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
