% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_builders.R
\name{plotBuilder_lcc}
\alias{plotBuilder_lcc}
\title{Generate a Longitudinal Concordance Correlation Plot}
\usage{
plotBuilder_lcc(rho, ENV.LCC, tk.plot, CCC, tk.plot2, ldb, model, ci, arg, ...)
}
\arguments{
\item{rho}{Vector of LCC values.}

\item{ENV.LCC}{Environment matrix for LCC values, used for confidence intervals.}

\item{tk.plot}{Time points for LCC values.}

\item{CCC}{Matrix or list of CCC values.}

\item{tk.plot2}{Time points for CCC values.}

\item{ldb}{Number of levels in the data.}

\item{model}{The model object from which data was extracted.}

\item{ci}{Logical, indicating if confidence intervals should be included.}

\item{arg}{List of graphical arguments.}

\item{...}{Additional arguments for ggplot.}
}
\description{
Produces a longitudinal concordance correlation plot from fitted 
and sampled values with optional non-parametric confidence intervals.
}
\keyword{internal}
