% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBuilder_lpc.R
\name{plotBuilder_lpc}
\alias{plotBuilder_lpc}
\title{Internal Function to Produce a Longitudinal Pearson Correlation Plot}
\description{
Produces a longitudinal Pearson correlation plot from fitted 
  and sampled values, with optional non-parametric confidence intervals.
}
\details{
Returns an initial plot for the longitudinal Pearson correlation.
}
\author{
Thiago de Paula Oliveira, \email{thiago.paula.oliveira@alumni.usp.br}
}
\keyword{internal}
