% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lccSummary.R
\name{print.summary.lcc}
\alias{print.summary.lcc}
\title{Print the summary of an lcc object}
\usage{
\method{print}{summary.lcc}(x, type, verbose = FALSE, digits = NULL,
  ...)
}
\arguments{
\item{x}{an object inheriting from class
\code{\link[lcc]{lcc}}, representing a fitted longitudinal
concordance correlation function.}

\item{type}{an optional character string specifying the type of
output to be returned. If \code{type="model"}, prints the summary
of the polynomial mixed-effects regression model. If
\code{type="lcc"}, prints the summary of the fitted and sampled
values for LCC, LPC, and LA as well as the concordance correlation
coefficient between fitted LCC values and observed values as
goodness of fit (gof) measurement. Defaults to \code{type="lcc"}.}

\item{verbose}{an optional logical value used to control the amount
of printed output when \code{type = "model"}. Defaults to
\code{FALSE}}

\item{digits}{a non-null value for \code{digits} specifies the
minimum number of significant digits to be printed in values. The
default, \code{NULL}.}

\item{...}{not used.}
}
\description{
Information summarizing the fitted longitudinal
  concordance correlation is printed. This includes the AIC, BIC,
  and log-likelihood at convergence. If \code{type = "lcc"}, prints
  the fitted values while \code{type = "model"} prints the fixed
  effects estimates and their standard errors, standard deviations,
  correlations for the random effects, within-group correlation, and
  variance function parameters.
}
\examples{

data(hue)
## Second degree polynomial model with random intercept, slope and
## quadratic term
fm1<-lcc(dataset = hue, subject = "Fruit", resp = "H_mean",
         method = "Method", time = "Time", qf = 2, qr = 2)
print(summary(fm1, type="model"))
}
\seealso{
\code{\link{summary.lcc}}, \code{\link{lccPlot}},
  \code{\link[lcc]{lcc}}
}
\author{
Thiago de Paula Oliveira,
  \email{thiago.paula.oliveira@usp.br}
}
